Update the reset password function, routes, html + removing files that are no longer viable
