--- forms.py
@@ -1,95 +1,0 @@
-from flask_wtf import FlaskForm
-from flask_wtf.file import FileField, FileAllowed
-from flask_login import current_user
-from wtforms import StringField, PasswordField, SubmitField, BooleanField
-from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
-from models import *
-
-
-class RegistrationForm(FlaskForm):
-    username = StringField(
-        "Username", validators=[DataRequired(), Length(min=2, max=20)]
-    )
-    email = StringField("Email", validators=[DataRequired(), Email()])
-    password = PasswordField("Password", validators=[DataRequired()])
-    confirm_password = PasswordField(
-        "Confirm Password", validators=[DataRequired(), EqualTo("password")]
-    )
-    submit = SubmitField("Sign Up")
-
-    @staticmethod
-    def validate_username(username):
-        user = Userdb.query.filter_by(username=username.data).first()
-        if user:
-            raise ValidationError(
-                "That username is taken. Please choose a different one."
-            )
-
-    @staticmethod
-    def validate_email(email):
-        user = Userdb.query.filter_by(email=email.data).first()
-        if user:
-            raise ValidationError("That email is taken. Please choose a different one.")
-
-
-class LoginForm(FlaskForm):
-    email = StringField("Email", validators=[DataRequired(), Email()])
-    password = PasswordField("Password", validators=[DataRequired()])
-    remember = BooleanField("Remember Me")
-    submit = SubmitField("Login")
-
-
-class ExpensesForm(FlaskForm):
-    expense = StringField("Expense", validators=[DataRequired()])
-    price = StringField("Price", validaters=[DataRequired()])
-    add = SubmitField("Add")
-
-
-class UpdateAccountForm(FlaskForm):
-    username = StringField(
-        "Username", validators=[DataRequired(), Length(min=2, max=20)]
-    )
-    email = StringField("Email", validators=[DataRequired(), Email()])
-    picture = FileField(
-        "Update Profile Picture", validators=[FileAllowed(["jpg", "png"])]
-    )
-    submit = SubmitField("Update")
-
-    @staticmethod
-    def validate_username(username):
-        if username.data != current_user.username:
-            user = Userdb.query.filter_by(username=username.data).first()
-            if user:
-                raise ValidationError(
-                    "That username is taken. Please choose a different one."
-                )
-
-    @staticmethod
-    def validate_email(email):
-        if email.data != current_user.email:
-            user = User.query.filter_by(email=email.data).first()
-            if user:
-                raise ValidationError(
-                    "That email is taken. Please choose a different one."
-                )
-
-
-class RequestResetForm(FlaskForm):
-    email = StringField("Email", validators=[DataRequired(), Email()])
-    submit = SubmitField("Request Password Reset")
-
-    @staticmethod
-    def validate_email(email):
-        user = Userdb.query.filter_by(email=email.data).first()
-        if user is None:
-            raise ValidationError(
-                "There is no account with that email. You must register first."
-            )
-
-
-class ResetPasswordForm(FlaskForm):
-    password = PasswordField("Password", validators=[DataRequired()])
-    confirm_password = PasswordField(
-        "Confirm Password", validators=[DataRequired(), EqualTo("password")]
-    )
-    submit = SubmitField("Reset Password")
