--- models.py
@@ -1,51 +1,0 @@
-from routes import login_manager
-from routes import db
-from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
-from flask_login import UserMixin
-
-
-@login_manager.user_loader
-def load_user(user_id):
-    return Userdb.query.get(int(user_id))
-
-
-class Userdb(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(120), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    image_file = db.Column(db.String(20), nullable=False, default="default.jpg")
-    password = db.Column(db.String(60), nullable=False)
-
-    def get_reset_token(self, expires_sec=1800):
-        s = Serializer(app.config["SECRET_KEY"], expires_sec)
-        return s.dumps({"user_id": self.id}).decode("utf-8")
-
-    @staticmethod
-    def verify_reset_token(token):
-        s = Serializer(app.config["SECRET_KEY"])
-        try:
-            user_id = s.loads(token)["user_id"]
-        except ValueError:
-            return None
-        return Userdb.query.get(user_id)
-
-    def __repr__(self):
-        return f"User('{self.username}', '{self.email}', '{self.image_file}')"
-
-
-class Expensedb(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    expense = db.Column(db.Integer, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("userdb.id"), nullable=False)
-
-    def __repr__(self):
-        return "<Expensedb %r>" % self.expense
-
-
-class Budgetdb(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    budget = db.Column(db.Integer, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("userdb.id"), nullable=False)
-
-    def __repr__(self):
-        return "<Artistdb %r>" % self.budget
