--- static/profile_pics/avatar-default-icon.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://icons.iconarchive.com/icons/papirus-team/papirus-status/512/avatar-default-icon.png
-HostUrl=https://icons.iconarchive.com/icons/papirus-team/papirus-status/512/avatar-default-icon.png
